Add BinaryParser class for encoding/decoding framework
This commit adds a new BinaryParser class to handle encoding and decoding tasks. It also includes changes in CryptoHelperTests.swift for improved key pair generation, and updates to the main project configuration to include the BinaryParser.swift file in project sources. Finally, it removes duplicative coding from the NanoTDF.swift file to make use of the newly-added BinaryParser class.
